TCP: Process incoming data even when some buffers do not end with newline
TCP does segmentation as it pleases, and there is no guarantee that segments
would end with a newline.  Sometimes we may receive multiple segments which
contain many packets, and some packets are split between multiple TCP segments
and incoming buffers, and we should just process the data instead of buffer
it all and decode only when a buffer ends with a newline. The old behaviour
may block processing for quite a while when a lot of data comes in.